+++ packages/ratel/lib/src/runtime/socket_definition.dart
@@ -5,10 +5,14 @@ final class SocketDefinition<C> {
   const SocketDefinition({
     required this.path,
     required this.invoke,
+    this.isProtected = false,
+    this.requiredRoles = const [],
     this.parameters = const [],
   });
 
   final String path;
+  final bool isProtected;
+  final List<String> requiredRoles;
   final List<RouteParameter> parameters;
   final ControllerInvoker<C> invoke;
 }